Removing the int casting

+++ Microsoft.Azure.Cosmos/src/Handler/TransportHandler.cs
@@ -127,7 +127,7 @@ namespace Microsoft.Azure.Cosmos.Handlers
                     if (dce.StatusCode == System.Net.HttpStatusCode.Unauthorized || dce.StatusCode == System.Net.HttpStatusCode.Forbidden)
                     {
                         TimeSpan authProvideLifeSpan = ((AuthorizationTokenProvider)this.client.DocumentClient.cosmosAuthorization).GetAge();
-                        processMessageAsyncTrace.AddDatum("AuthProvider LifeSpan InSec", (int)authProvideLifeSpan.TotalSeconds);
+                        processMessageAsyncTrace.AddDatum("AuthProvider LifeSpan InSec", authProvideLifeSpan.TotalSeconds);
                     }
 
                     throw;
